+++ setup.cfg.old
@@ -4,7 +4,7 @@ with-spec=1
 spec-color=1
 with-coverage=1
 cover-erase=1
-cover-package=models
+cover-package=source
 
 [coverage:report]
 show_missing = True
