--- tests/test_temp.py
@@ -1,30 +1,0 @@
-"""
-Template Test Cases, please update
-"""
-from unittest import TestCase
-from source import *
-
-class TestTemplate(TestCase):
-    """ Test Template, please update """
-
-    @classmethod
-    def setUpClass(cls):
-        """ Do something to set up tests, please update """
-
-    @classmethod
-    def tearDownClass(cls):
-        """ Do something to clean up tests, please update """
-
-    def setUp(self):
-        """ Ensure each test runs clean, please update """
-
-    def tearDown(self):
-        """ Ensure each test runs clean, please update """
-
-    ###########################################################################
-    #   T E S T  C A S E S
-    ###########################################################################
-
-    def test_template(self):
-        """ Template test case, please update """
-        self.assertEqual(True,False)
